Fix indirect jump bug due to bad offset
This bug was in the extraction code (which was fixed). We also fix a
bug in the that calculated the wrong auipc base (because we made a false
precondition on what the base/base' values mean).

The indirect jump bug was due to a simple arithmetic error, though
because there were no proofs determining the validity of the table
values, this bug was left unpatched.

Note that in the next commit we use base/4 values, since that was our
initial assumption that was made.