Add new induction on invariants ltac code!
This adds a induction tactic for invariant type which will automatically
locate preconditions that reference the initial state and try to
autogenerate corresponding preconditions after some execution occurred.

This requires a bit of ugly ltac code with a separate library for
looping through all the premises in ltac (is there a better way to do
this?) Maybe some of this code can simplified if we shelved the subgoals for
the preconditions after some executions for some moment and then revisit
them later on, instead of trying to weave through those subgoals, using
some ungodly amount of recursive tactics?

Also reprove/simplifies the strcmp proofs by using this induction tactic
instead.