Better invariant induction tactic
Rewrites premises that have been reverted to match out the non-initial
state. We also clear out any premises in the goal referring to the
initial state, but warn the user that we are doing this. We probably
don't want any premises that involve the initial state because we can't
really easily reason with that after we destruct all invariant points.

Instead the user probably meant to replace it with some invariant over
some set of locations, which we also support by a special
`invariant <inv> from <pre>` tactic.

Removes the Looping library too since we don't use it anymore.

Finally, updates all the sample proofs to contain our new invariant
induction tactic.