Add functional interpretation for exp -> eval_exp.
Finterp currently only does reasoning about interpretation of already
executed entities (i.e. given some execution of a statement/evaluation
of a expression, give the value), but we might also like to evaluate
values from just an expression in the determininstic case, i.e. if the
expression does not contain unknowns. This allows us to evaluate the
expression deltas as computed from a function, i.e. when tracing a
function.

This adds a number of required proofs, type preservations, and other
misc stuff to make evaluation of deterministic expressions sound. Note
that the operation is not complete due to the presence of certain
nondeterminism via the unknowns.