Restructure reachablility hints.
This specifies what types of hints the user would have to actually give.
We narrow up the hints to just have the user give jump address locations
for each possible expressions (or a "symbolic" location if the user is
sure that the code jumps to a non-program area). This also puts less
burden on the user to prove their hints' correctness.

We also merge the reachability and tracing computation into one single
function (since we now have a binary tree data structure to query known
reachable addresses) to simplify some of the code. By doing this, the
user also has the ability to query the store deltas of an address to
query the known current CPU state at that expression/instruction/address.

Finally, we added more modular arithmetic proofs and to help us prove
memory preservation. This creates a new type of memory-frame
preservation with the goal proving the minimum absolute distance away
from some target address. Doing so allows us to not have to destruct on
whether an address access overflows the address space (allowing us to
reduce the number of preconditions).